refactor: substitute sql native query with SPRING DATA JPA internal method

+++ src/main/java/dev/be/sns/controller/PostController.java
@@ -57,8 +57,8 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication){
-        int count = postService.likeCount(postId);
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication){
+        long count = postService.likeCount(postId);
         return Response.success(count);
     }
